manual_demo: fix HTTP route policy
The existing demo is broken because of incorrect
host header matching specified which prevents
bookbuyer from accessing the bookstore service.
The host header in the HTTP request includes
the port, but the spec was missing this. Since
the host header feature is not complete (see
openservicemesh/osm#2369) and this is unnecessary
for this demo, this change removes it.

The demo works as expected with SMI policies
with this fix.

Resolves openservicemesh/osm#3491

Signed-off-by: Shashank Ram <shashr2204@gmail.com>