ingress: document IngressBackend with Contour
Documents using the IngressBackend API for ingress with
Contour ingress controller. Adds a section for BYO ingress
controller and removes outdated feature and demo docs.
Demos with other ingress controllers (Nginx, AGIC etc.)
are not any different when it comes to using the IngressBackend
API, as demoed with Contour already.

Also addresses duplicate indexing of topics for
traffic_management.

Part of:
- openservicemesh/osm#3779
- openservicemesh/osm#2365
- openservicemesh/osm#3700

Signed-off-by: Shashank Ram <shashr2204@gmail.com>