config: clarify latest version tag on website
Clarifies that the latest version tag on the website
refers to an unreleased version. Without this clarification
users are misled to think the latest version is a released
version. Although there is a banner on the release specific
docs page, most users directly click on the "latest" version
without ever going to the release specific page.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>