docs(demo): separate root traffic split service from backend
These changes update the manual demo to use the more standard method of
traffic splitting, where the root service is not also a backend service.

To do this, the current `bookstore` service has been renamed to
`bookstore-v1` and a new `bookstore` service will serve as the root for
the TrafficSplit.

The only change in the overall steps of the demo is a new step to update
the `BOOKSTORE_SVC` environment variable in the `bookbuyer` and
`bookthief` from `bookstore-v1` to `bookstore` right before creating the
first TrafficSplit.

Signed-off-by: Jon Huhn <johuhn@microsoft.com>