docs(metrics): clarify instructions for enabling metrics scraping
Clarifies in the observability section of the getting started
guide that metrics scraping must be enabled on a namespace for the
pods in that namespace to be annotated by the osm injector. Also
adds this clarification to the metrics how to guide intro section.

Removes incorrect statement about pods needing to be recreated for
the annotations to be to added or removed corresponding to
metrics being enabled or disabled on the namespace.

Signed-off-by: jaellio <jaellio@microsoft.com>