inline port-forward scripts in demo
The existing port-forward-all.sh script in the code repo forwards ports
for the demo described here and the automated demo defined in the code
repo. The automated demo defines a `bookstore-v1` deployment where the
demo here names it only `bookstore`, both of which are port-forwarded by
the script, which then throws nondescript messages related to either one
of those not existing when running either demo.

These changes inline the specific scripts needed at each point in the
demo to ensure only components that are expected to exist have ports
forwarded so no error messages are expected.

Part of https://github.com/openservicemesh/osm/issues/3521

Signed-off-by: Jon Huhn <johuhn@microsoft.com>