docs/demo: define global rate limit policies at the beginning (#447)
Updates the global rate limit demo to define the external
rate limit policies in the ConfigMap at the beginning versus
at every step in the demo. When the rate limit ConfigMap
is updated at different stages in the demo, the timing
of when the rate limit config is picked up by the RLS
service is determined by when kubelet refreshes the
ConfigMap changes onto the pods. This results in a
non-deterministic demo experience.

With this change, similar to the e2e test, all the
policies applicable to the RLS are applied at the
beginning, so that the demo experience is smoother
and not driven by when kubelet syncs the ConfigMap
changes onto the RLS pod.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>